pyproject.toml
[project]
name = "aura-snowflake-playground"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "notebook>=7.4.7",
    "python-dotenv>=1.1.1",
    "snowflake>=1.8.0",
    "snowflake-connector-python>=3.18.0",
    "snowflake-ml-python>=1.16.0",
    "snowflake-snowpark-python>=1.40.0",
    "streamlit>=1.50.0",
]
